storage account modularization

+++ azure-toolkit-libs/azure-toolkit-storage-lib/src/main/java/com/microsoft/azure/toolkit/lib/storage/service/StorageAccount.java
@@ -7,8 +7,10 @@ package com.microsoft.azure.toolkit.lib.storage.service;
 
 import com.azure.core.management.exception.ManagementException;
 import com.azure.resourcemanager.storage.StorageManager;
+import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.common.entity.AbstractAzureResource;
 import com.microsoft.azure.toolkit.lib.common.entity.IAzureResource;
+import com.microsoft.azure.toolkit.lib.common.entity.Removable;
 import com.microsoft.azure.toolkit.lib.common.event.AzureOperationEvent;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.storage.model.StorageAccountEntity;
@@ -17,13 +19,13 @@ import org.apache.http.HttpStatus;
 import javax.annotation.Nonnull;
 
 public class StorageAccount extends AbstractAzureResource<StorageAccount, StorageAccountEntity, com.azure.resourcemanager.storage.models.StorageAccount>
-        implements AzureOperationEvent.Source<StorageAccount>, IAzureResource<StorageAccountEntity> {
+        implements Removable, AzureOperationEvent.Source<StorageAccount>, IAzureResource<StorageAccountEntity> {
     @Nonnull
     private final StorageManager manager;
 
-    public StorageAccount(@Nonnull StorageManager manager, @Nonnull com.azure.resourcemanager.storage.models.StorageAccount server) {
+    public StorageAccount(@Nonnull com.azure.resourcemanager.storage.models.StorageAccount server) {
         super(new StorageAccountEntity(server));
-        this.manager = manager;
+        this.manager = server.manager();
     }
 
     @Override
@@ -42,6 +44,15 @@ public class StorageAccount extends AbstractAzureResource<StorageAccount, Storag
 
     @AzureOperation(name = "storage|account.delete", params = {"this.entity().getName()"}, type = AzureOperation.Type.SERVICE)
     public void delete() {
-        manager.storageAccounts().deleteById(this.entity.getId());
+        if (this.exists()) {
+            this.status(Status.PENDING);
+            manager.storageAccounts().deleteById(this.entity.getId());
+            Azure.az(AzureStorageAccount.class).refresh();
+        }
+    }
+
+    @Override
+    public void remove() {
+        this.delete();
     }
 }
